Reorganize directory.
