+++ cdk/src/open-data-platform/data-plane/data-plane-stack.ts
@@ -5,10 +5,9 @@ import * as ec2 from 'aws-cdk-lib/aws-ec2';
 import * as rds from 'aws-cdk-lib/aws-rds';
 import { Construct } from 'constructs';
 import { CommonProps } from '../../util';
-import { Schema } from './schema';
-import {DataImportStack} from '../lambda/data-import-stack';
-import {DatabaseUserCredentials} from './db-user-credentials';
-import {ConnectionString} from './connection-string';
+import { Schema } from './schema/schema';
+import { DataImportStack } from '../lambda/data-import-stack';
+import { DatabaseUserCredentials } from './db-user-credentials/db-user-credentials';
 
 interface DataPlaneProps extends CommonProps {
   vpc: ec2.IVpc;
@@ -17,7 +16,6 @@ interface DataPlaneProps extends CommonProps {
 export class DataPlaneStack extends Stack {
   readonly cluster: rds.ServerlessCluster;
   readonly tileserverCredentials: DatabaseUserCredentials;
-  readonly tileserverConnectionString: ConnectionString;
 
   constructor(scope: Construct, id: string, props: DataPlaneProps) {
     super(scope, id, props);
@@ -32,9 +30,9 @@ export class DataPlaneStack extends Stack {
     this.cluster = new rds.ServerlessCluster(this, 'MainCluster', {
       engine: rds.DatabaseClusterEngine.AURORA_POSTGRESQL,
       parameterGroup: rds.ParameterGroup.fromParameterGroupName(
-          this,
-          'ParameterGroup',
-          'default.aurora-postgresql10',
+        this,
+        'ParameterGroup',
+        'default.aurora-postgresql10',
       ),
       vpc,
       vpcSubnets: {
@@ -51,35 +49,31 @@ export class DataPlaneStack extends Stack {
     });
 
     // default DB name.
-    const dbName = 'postgres';
+    const databaseName = 'postgres';
 
     // Credentials for the tile server to access the cluster.
-    this.tileserverCredentials = new DatabaseUserCredentials(this, 'MainClusterMartinTileserverCredentials', {
+    this.tileserverCredentials = new DatabaseUserCredentials(this, 'TileserverCredentials', {
       cluster: this.cluster,
-      username: 'tileserver'
-    })
-
-    new ConnectionString(this, 'MartinTileserverConnectionString', {
-      sourceSecret: this.tileserverCredentials.secret,
-      databaseName: dbName
-    })
+      username: 'tileserver',
+      databaseName,
+    });
 
     // Initialize the DB with linked SQL file.
     const rootSchema = new Schema(this, 'RootSchema', {
       cluster: this.cluster,
       vpc,
-      db: dbName,
+      db: databaseName,
       schemaFileName: 'schema.sql',
       credentialsSecret: this.cluster.secret!,
-      userCredentials: [this.tileserverCredentials.secret]
+      // userCredentials: [this.tileserverCredentials.secret],
     });
-    rootSchema.node.addDependency(this.tileserverCredentials.secret);
+    // rootSchema.node.addDependency(this.tileserverCredentials.secret);
 
     new DataImportStack(this, 'DataImportStack', {
       cluster: this.cluster,
       vpc: vpc,
-      db: dbName,
+      db: databaseName,
       credentialsSecret: this.cluster.secret!,
-    })
+    });
   }
 }
