+++ cdk/src/open-data-platform/data-plane/schema/schema.sql
